+++ src/main/kotlin/com/dodo/dodolistserver/common/exception/auth/FobiddenException.kt
@@ -1,4 +1,4 @@
-package com.dodo.dodolistserver.common.exception
+package com.dodo.dodolistserver.common.exception.auth
 
 private const val FORIBIDDEN_EXCECPTION_MSG = "허용되지 않은 사용자 입니다."
