+++ src/main/kotlin/com/dodo/dodolistserver/common/exception/auth/TokenExpireException.kt
@@ -1,4 +1,4 @@
-package com.dodo.dodolistserver.common.exception
+package com.dodo.dodolistserver.common.exception.auth
 
 private const val TOKEN_EXPIRE_ERR_MSG = "Token이 만료가 되었습니다."
