+++ src/main/kotlin/com/dodo/dodolistserver/common/exception/auth/TokenValidationException.kt
@@ -1,4 +1,4 @@
-package com.dodo.dodolistserver.common.exception
+package com.dodo.dodolistserver.common.exception.auth
 
 private const val TOKEN_VALIDATION_ERR_MSG: String = "Token이 validate하지 않습니다."
