+++ src/main/kotlin/com/dodo/dodolistserver/common/exception/auth/UserNotFoundException.kt
@@ -1,4 +1,6 @@
-package com.dodo.dodolistserver.common.exception
+package com.dodo.dodolistserver.common.exception.auth
+
+import com.dodo.dodolistserver.common.exception.NotFoundException
 
 private const val USER_NOT_FOUND_EXCEPTION_MSG = "등록된 회원이 아닙니다."
